refactor: change the way adding src tvm id in access log
```
void Middleware::OnCallStart(MiddlewareCallContext& context) const {
    // Add common storage to extra log tags for other middlewares
    context.GetStorageContext().Emplace(kLogExtraTag, logging::LogExtra{});
}
```

This is a rather difficult operation involving dynamic allocation. It is cheaper not to create a kLogExtraTag key if it is not used.

On the other hand `SetAdditionalLogKeys` is a bit more convinient contract, cause it is not neccessary for a middleware check if LogExtra variable exists
commit_hash:5d13b5f73902e2b3fe4083cc47025728938bc0a9