feat userver-recentperiod: lower contention effects on buckets
Under high contention, `RecentPeriod` tended to increment the current index twice, which led to "zeroed" buckets that stopped the cycle in `GetStatsForPeriod`. In cases of a stable signal, this caused drops in the returned value down to 0, with restoration to the actual value only after `MaxDuration`—when all buckets had been revisited.

This fix does not aim to provide an algorithmically safe solution, but it does practically reduce the frequency of such drops.
commit_hash:ec2be57aec578dcce323ac1a17468c3e1779a091